Fix failing tests
A number of tests failed, because there's check in the LSMTool source code
on the version number of SciPy. This check breaks if the version number is
that of a release candidate. Now, I'm not sure why `python setup.py test`
pulls in a pre-release version of SciPy, but it easier to fix this in the
source code.

TODO: the use of `distutils.version` is deprecated; it shoud be replaced
      with `packaging.version`